requirements.txt
anyio==3.4.0
arrow==1.2.1
asgiref==3.4.1
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
fastapi==0.68.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.3
httptools==0.2.0
httpx==0.21.1
hyperframe==6.0.1
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
nb-cli==0.5.1
nonebot-adapter-cqhttp==2.0.0a16
nonebot2==2.0.0a16
poyo==0.5.0
prompt-toolkit==3.0.24
pydantic==1.8.2
pyfiglet==0.8.post1
pygtrie==2.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
PyYAML==6.0
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
text-unidecode==1.3
tomlkit==0.7.2
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1

peewee~=3.14.8
aiofiles~=0.8.0
Pillow~=8.4.0